requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
bcrypt
cryptography
pyjwt
